requirements.txt
asgiref==3.2.10
astroid==2.2.5
autopep8==1.4.4
boto3==1.20.45
botocore==1.23.45
certifi==2019.6.16
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.43.0
django-cors-headers==3.11.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.22.1
oauthlib==3.0.1
pep8==1.7.1
praat-parselmouth==0.4.1
psycopg2==2.9.2
pycodestyle==2.5.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.3.1
python-dateutil==2.8.2
python-dotenv==0.15.0
python3-openid==3.1.0
pytz==2019.1
requests==2.25.1
requests-oauthlib==1.2.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.3.0
urllib3==1.26.8
whitenoise==5.3.0
wrapt==1.11.2
